backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==2.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.11
asyncpg==0.30.0
sqlalchemy==2.0.30
argon2-cffi==23.1.0